Fix joystick map issue?

+++ include/input.h
@@ -1,28 +1,19 @@
 #pragma once
 #include "adc.h"
 #include <stdbool.h>
-typedef struct
-{
-    int8_t x;
-    int8_t y;
-    bool btn;
+typedef struct {
+  int x;
+  int y;
+  bool btn;
 
-    // uint8_t btn
+  // uint8_t btn
 } joystick_xy_t;
 
-typedef enum
-{
-    LEFT,
-    RIGHT,
-    UP,
-    DOWN,
-    NEUTRAL
-} joystick_dir_t;
+typedef enum { LEFT, RIGHT, UP, DOWN, NEUTRAL } joystick_dir_t;
 
-typedef struct
-{
-    int16_t x;
-    int16_t y;
+typedef struct {
+  int16_t x;
+  int16_t y;
 } touch_pad_xy_t;
 
 // typedef struct ADC
@@ -30,10 +21,9 @@ typedef struct
 //     struct joystick;
 //     struct touch_pad;
 // };
-typedef struct ADC
-{
-    joystick_xy_t joystick;
-    touch_pad_xy_t touch_pad;
+typedef struct ADC {
+  joystick_xy_t joystick;
+  touch_pad_xy_t touch_pad;
 } ADC;
 
 void map_joystick(joystick_xy_t *joystick);
